Give writers an error option for encode()
The python encode() function takes two arguments: a target encoding and
an optional method to use when the target encoding does not contain the
desired character.  unicodecsv's writer and DictWriter objects now take
an optional errors option which will be passed down to the underlying
encode().

The default errors option for unicodecsv is 'strict' which is also the
default for encode().