Python 3 fallback when no encoding given
`unicodecsv` is a drop-in replacement for the `csv` module on Python 2.
On Python 3, that model doesn't work as well, because of the strict
distinction between bytes and text.

However, it still surprises people that `unicodecsv` isn't a drop-in
replacement for Python 3 as well.

To make this possible, on Python 3 `unicodecsv` will fall back to the
bult-in `csv` module when no `encoding` argument is given. This allows
code written for Python 3's csv module to work without change.

However, it comes with a drawback: If you're relying on the default
utf-8 encoding of unicodecsv in Python 2, you'll get stronge,
hard-to-decipher errors on Python 3.